+++ docs/CHANGELOG.md
@@ -6,6 +6,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.0.24] - unreleased
+
+### Added
+
+- Placeholder `{timestamp}` in the download output directory path, replaced with the
+  current UTC time
+
+### Changed
+
+- Commands `queue download` and `playlist download` no longer create a timestamped
+  directory inside the output directory
+- The download commands create the output directory if missing
+
+
 ## [0.0.23] - 2026-07-28
 
 ### Added